requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.1
cryptography==42.0.2
python-jose==3.3.0

bcrypt==4.1.2
pydantic[email]==2.6.1
python-multipart==0.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]
python-multipart